requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
importlib-metadata==3.7.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
pycodestyle==2.7.0
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.1
